fix: move simple utils to global code
