+++ src/bkuser_global/utils.py
@@ -20,11 +20,14 @@ def get_timezone_offset():
     return datetime.timedelta(seconds=offset_seconds)
 
 
-def force_str_2_bool(bool_str, raise_if_unknown=False):
+def force_str_2_bool(bool_str: str, raise_if_unknown: bool = False) -> bool:
     """convent 'True' or 'False' to bool, using on query_param"""
-    if bool_str in ["True", "true"]:
+    if isinstance(bool_str, bool):
+        return bool_str
+
+    if bool_str in ["True", "true", "1"]:
         return True
-    elif bool_str in ["False", "false"]:
+    elif bool_str in ["False", "false", "0"]:
         return False
 
     if raise_if_unknown:
